+++ modules/crypto/src/encoding.cpp
@@ -9,8 +9,8 @@ namespace lcxx::encode {
     {
         // Add additional 4 bytes as buffer, because the rounding would just make it more complicated.
         std::vector< unsigned char > encoded( bytes.size() * 4 / 3 + 4 + 1, 0 );
-        auto                         true_size =
-            EVP_EncodeBlock( encoded.data(), reinterpret_cast< unsigned char const * >( bytes.data() ), bytes.size() );
+        auto true_size = EVP_EncodeBlock( encoded.data(), reinterpret_cast< unsigned char const * >( bytes.data() ),
+                                          static_cast< int >( bytes.size() ) );
 
         // Do not include terminating null-character
         return std::string{ encoded.begin(), encoded.begin() + true_size };
@@ -26,7 +26,8 @@ namespace lcxx::decode {
         std::vector< std::byte > decoded( input.size() * 3 / 4, std::byte{ 0 } );
 
         EVP_DecodeBlock( reinterpret_cast< unsigned char * >( decoded.data() ),
-                         reinterpret_cast< unsigned char const * >( input.data() ), input.size() );
+                         reinterpret_cast< unsigned char const * >( input.data() ),
+                         static_cast< int >( input.size() ) );
 
         // Remove empty bytes at the end
         if ( input.size() >= 1 && input.back() == '=' )
